DNSSEC-secured host-to-realm lookups with libunbound
For host-to-realm lookups, we cannot rely on the Kerberos protocol to validate
the connection found.  For that reason, such lookups are not considered
trustworthy, and generally not advised.  Using Unbound in a security-enforcing
mode, with no iteration up the DNS tree, we obtain a secure variation that
can be used for looking up realms for host names in DNS.

Unbound is not needed for SRV lookups leading to a KDC; for this use, there is
generally a shared key that can be used to validate the location found.  These
can safely be done over unprotected DNS.  The two can be combined; when this is
done, Unbound will take over from unprotected DNS for the _kerberos TXT lookups
of host-to-realm (and domain-to-realm) translations.

Read more in draft-vanrein-dnstxt-krb1